+++ src/accel/SetupOptions.cc
@@ -12,9 +12,7 @@
 #include "corecel/math/ArrayUtils.hh"
 #include "geocel/GeantGeoUtils.hh"
 #include "geocel/GeantUtils.hh"
-#if CELERITAS_USE_COVFIE
-#    include "celeritas/field/CartMapFieldInput.hh"
-#endif
+#include "celeritas/field/CartMapFieldInput.hh"
 #include "celeritas/field/CylMapFieldInput.hh"
 #include "celeritas/field/RZMapFieldInput.hh"
 #include "celeritas/field/UniformFieldData.hh"
@@ -169,13 +167,11 @@ void ProblemSetup::operator()(inp::Problem& p) const
         CELER_LOG(debug) << "Getting Cyl map field";
         p.field = u->get_field();
     }
-#if CELERITAS_USE_COVFIE
     else if (auto* u = so.make_along_step.target<CartMapFieldAlongStepFactory>())
     {
         CELER_LOG(debug) << "Getting Cyl map field";
         p.field = u->get_field();
     }
-#endif
     else
     {
         CELER_NOT_IMPLEMENTED("processing generic along-step factory");
