+++ src/celeritas/setup/Problem.cc
@@ -27,9 +27,7 @@
 #include "geocel/GeantGdmlLoader.hh"
 #include "celeritas/Quantities.hh"
 #include "celeritas/Types.hh"
-#if CELERITAS_USE_COVFIE
-#    include "celeritas/alongstep/AlongStepCartMapFieldMscAction.hh"
-#endif
+#include "celeritas/alongstep/AlongStepCartMapFieldMscAction.hh"
 #include "celeritas/alongstep/AlongStepCylMapFieldMscAction.hh"
 #include "celeritas/alongstep/AlongStepGeneralLinearAction.hh"
 #include "celeritas/alongstep/AlongStepRZMapFieldMscAction.hh"
@@ -317,7 +315,6 @@ auto build_along_step(inp::Field const& var_field,
                                         msc,
                                         eloss);
             },
-#if CELERITAS_USE_COVFIE
             [&](inp::CartMapField const& field) {
                 using ASA = AlongStepCartMapFieldMscAction;
                 return ASA::from_params(next_id,
@@ -327,7 +324,6 @@ auto build_along_step(inp::Field const& var_field,
                                         msc,
                                         eloss);
             },
-#endif
         }),
         var_field);
 }
